refactor(codegen): extract buildPropertiesForEvent to parsers-commons
