+++ packages/react-native-codegen/src/parsers/flow/components/events.js
@@ -22,7 +22,10 @@ const {
   throwIfBubblingTypeIsNull,
   throwIfArgumentPropsAreNull,
 } = require('../../error-utils');
-const {getEventArgument} = require('../../parsers-commons');
+const {
+  getEventArgument,
+  buildPropertiesForEvent,
+} = require('../../parsers-commons');
 const {
   emitBoolProp,
   emitDoubleProp,
@@ -67,7 +70,7 @@ function getPropertyType(
         typeAnnotation: {
           type: 'ObjectTypeAnnotation',
           properties: typeAnnotation.properties.map(member =>
-            buildPropertiesForEvent(member, parser),
+            buildPropertiesForEvent(member, parser, getPropertyType),
           ),
         },
       };
@@ -126,7 +129,7 @@ function extractArrayElementType(
       return {
         type: 'ObjectTypeAnnotation',
         properties: typeAnnotation.properties.map(member =>
-          buildPropertiesForEvent(member, parser),
+          buildPropertiesForEvent(member, parser, getPropertyType),
         ),
       };
     case 'ArrayTypeAnnotation':
@@ -229,19 +232,6 @@ function findEventArgumentsAndType(
   }
 }
 
-function buildPropertiesForEvent(
-  /* $FlowFixMe[missing-local-annot] The type annotation(s) required by Flow's
-   * LTI update could not be added via codemod */
-  property,
-  parser: Parser,
-): NamedShape<EventTypeAnnotation> {
-  const name = property.key.name;
-  const optional = parser.isOptionalProperty(property);
-  const typeAnnotation = parser.getTypeAnnotationFromProperty(property);
-
-  return getPropertyType(name, optional, typeAnnotation, parser);
-}
-
 function buildEventSchema(
   types: TypeMap,
   property: EventTypeAST,
@@ -283,8 +273,8 @@ function buildEventSchema(
         type: 'EventTypeAnnotation',
         argument: getEventArgument(
           nonNullableArgumentProps,
-          buildPropertiesForEvent,
           parser,
+          getPropertyType,
         ),
       },
     };
@@ -298,8 +288,8 @@ function buildEventSchema(
       type: 'EventTypeAnnotation',
       argument: getEventArgument(
         nonNullableArgumentProps,
-        buildPropertiesForEvent,
         parser,
+        getPropertyType,
       ),
     },
   };
